+++ src/Module/System/Update/Versions.php
@@ -362,6 +362,7 @@ use Ampache\Module\System\Update\Migration\V8\Migration800031;
 use Ampache\Module\System\Update\Migration\V8\Migration800032;
 use Ampache\Module\System\Update\Migration\V8\Migration800033;
 use Ampache\Module\System\Update\Migration\V8\Migration800034;
+use Ampache\Module\System\Update\Migration\V8\Migration800035;
 use Generator;
 
 /**
@@ -369,7 +370,7 @@ use Generator;
  */
 final class Versions
 {
-    public const int MAXIMUM_UPDATABLE_VERSION = 800034; // AMPACHE_VERSION (db_version)
+    public const int MAXIMUM_UPDATABLE_VERSION = 800035; // AMPACHE_VERSION (db_version)
 
     /** @var array<int, class-string<MigrationInterface>> List of available migrations */
     private static array $versions = [
@@ -710,6 +711,7 @@ final class Versions
         800032 => Migration800032::class,
         800033 => Migration800033::class,
         800034 => Migration800034::class,
+        800035 => Migration800035::class,
     ];
 
     /**